Fix compilation error in cam24i_sa()
git-svn-id: svn+ssh://savannah.psi.ch/afs/psi.ch/project/meg/svn/midas/trunk@4456 050218f5-8902-0410-8d0e-8a15d521e4f2

+++ drivers/camac/ces8210.c
@@ -167,7 +167,7 @@ INLINE void cam16i_sa(const int c, const int n, const int a, const int f,
 
 /*--input--------------------------------------------------------*/
 INLINE void cam24i_sa(const int c, const int n, const int a, const int f,
-                      unsigned long **d, const int r)
+                      DWORD **d, const int r)
 {
    int aa;
